requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
boto3==1.26.79
botocore==1.29.79
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
Django==4.1.5
django_debug_toolbar==3.8.1
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
jsonschema==4.17.3
msgpack==1.0.4
ocpp==0.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
s3transfer==0.6.0
service-identity==21.1.0
setuptools==65.6.3
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
websockets==10.4
wheel==0.38.4
zope.interface==5.5.2